pyproject.toml
[tool.poetry]
name = "spirvsmith"
version = "0.1.0"
description = ""
authors = ["Rayan Hatout <rayan.hatout@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
shortuuid = "^1.0.8"
hydra-core = "^1.1.1"
ulid-py = "^1.1"
pytest = "^7.1.1"
datadog = "^0.44.0"
daiquiri = "^3.0.1"
coverage = "^6.3.2"
flake8 = "^4.0.1"
Flask = "^2.1.1"
google-cloud-pubsub = "^2.12.0"
google-cloud-storage = "^2.2.1"
paramiko = "^2.10.3"
rsa = "^4.8"
yq = "^2.14.0"
google-cloud-bigquery = "^3.0.1"
firebase-admin = "^5.2.0"
ddtrace = "^1.0.1"
GPUtil = "^1.4.0"
py-cpuinfo = "^8.0.0"
GitPython = "^3.1.27"

[tool.poetry.dev-dependencies]
black = {version = "^22.3", allow-prereleases = true}
autoflake = "^1.4"
pytest-flakefinder = "^1.0.0"
pre-commit = "^2.18.1"
mypy = "^0.950"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
